implemented new endpoint to each action of noteController. Implemented handleStatusXXX.js

+++ controllers/noteController.js
@@ -1,62 +1,21 @@
 import { prisma } from "../db";
 import yn from "yn";
-import logger from "../services/logger";
 
-const noteController = async (params) => {
-  logger.info(
-    "controllers/noteController.js -> params = " +
-      JSON.stringify(params, null, 2)
-  );
+import {
+  handleStatus400,
+  handleStatus401,
+  handleStatus403,
+  handleStatus404,
+  handleStatus500,
+} from "../services/api/handleStatusXXX";
 
-  switch (params.action) {
-    case "findAll":
-      return findAll(params);
-    case "create":
-      return create(params);
-    case "update":
-      return update(params);
-    case "deleteById":
-      return deleteById(params.id);
-    case "archive":
-      return archive(params);
-    default:
-      return methodDoesNotExist();
-  }
-};
+const fileName = "controllers/noteController";
 
-const handleException = (method, message) => {
-  const errorMessage =
-    "controllers/noteController.js -> Exception in method " +
-    method +
-    ": " +
-    message;
-  logger.error(errorMessage);
-
-  return {
-    status: 500,
-    success: false,
-    data: [],
-    errorMessage: errorMessage,
-  };
-};
-
-const methodDoesNotExist = () => {
-  const errorMessage =
-    "controllers/noteController.js -> the action requested does not exist";
-  logger.info(errorMessage);
-
-  return {
-    status: 400,
-    success: false,
-    data: [],
-    errorMessage: errorMessage,
-  };
-};
-
-//* ----- Methods -----
-
-const findAll = async (params) => {
+export const findAll = async (params) => {
   try {
+    const paramsError = await checkParams(params, "findAll");
+    if (paramsError) return paramsError;
+
     const data = await prisma.Note.findMany({
       where: {
         archived: {
@@ -73,14 +32,17 @@ const findAll = async (params) => {
       errorMessage: null,
     };
   } catch (error) {
-    return handleException("findAll", error.message);
+    return handleStatus500(fileName + ".findAll()", error);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const create = async (params) => {
+export const create = async (params) => {
   try {
+    const paramsError = await checkParams(params, "create");
+    if (paramsError) return paramsError;
+
     const data = await prisma.Note.create({
       data: {
         title: params.title,
@@ -96,14 +58,18 @@ const create = async (params) => {
       errorMessage: null,
     };
   } catch (error) {
-    return handleException("create", error.message);
+    return handleStatus500(fileName + ".create()", error);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const update = async (params) => {
+export const update = async (params) => {
   try {
+    const paramsError = await checkParams(params, "update");
+    if (paramsError) return paramsError;
+
+    //update note
     const data = await prisma.Note.update({
       where: {
         id: Number(params.id),
@@ -121,14 +87,18 @@ const update = async (params) => {
       errorMessage: null,
     };
   } catch (error) {
-    return handleException("update", error.message);
+    return handleStatus500(fileName + ".update()", error);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const deleteById = async (id) => {
+export const deleteById = async (params) => {
   try {
+    const paramsError = await checkParams(params, "deleteById");
+    if (paramsError) return paramsError;
+
+    //delete note
     const data = await prisma.Note.delete({
       where: {
         id: Number(id),
@@ -141,17 +111,21 @@ const deleteById = async (id) => {
       errorMessage: null,
     };
   } catch (error) {
-    return handleException("deleteById", error.message);
+    return handleStatus500(fileName + ".deleteById()", error);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const archive = async (params) => {
+export const archive = async (params) => {
   try {
+    const paramsError = await checkParams(params, "archive");
+    if (paramsError) return paramsError;
+
+    //archive note
     const data = await prisma.Note.update({
       where: {
-        id: Number(params.id),
+        id: Number(id),
       },
       data: {
         archived: yn(params.archived),
@@ -165,10 +139,62 @@ const archive = async (params) => {
       errorMessage: null,
     };
   } catch (error) {
-    return handleException("archive", error.message);
+    return handleStatus500(fileName + ".archive()", error);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-export default noteController;
+const checkParams = async (params, foo) => {
+  //check for status 401: there no current user id
+  if (!Number(params.currentUserId)) {
+    return handleStatus401(
+      `${fileName}.${foo}()`,
+      "current user id is missing"
+    );
+  }
+
+  if (["create", "update"].includes(foo)) {
+    //check for status 400: params.title is required
+    if (!params.title?.trim()?.length) {
+      return handleStatus400(
+        `${fileName}.${foo}()`,
+        "title is required and must not be empty"
+      );
+    }
+  }
+
+  if (["update", "deleteById", "archive"].includes(foo)) {
+    //check for status 400: params.id is required and must be a number
+    if (!Number(params.id)) {
+      return handleStatus400(
+        `${fileName}.${foo}()`,
+        "id is required and must be a number"
+      );
+    }
+
+    const note = await prisma.Note.findUnique({
+      where: {
+        id: Number(params.id),
+      },
+    });
+
+    //check for status 404: note might not exist
+    if (!note) {
+      return handleStatus404(
+        `${fileName}.${foo}()`,
+        `note #${params.id} does not exist`
+      );
+    }
+
+    //check for status 403: the note does not belong to the current user
+    if (note.userId !== Number(params.currentUserId)) {
+      return handleStatus403(
+        `${fileName}.${foo}()`,
+        "the note trying to be accessed does not belong to the current user"
+      );
+    }
+  }
+
+  return null;
+};
